test(cli): split companion tests
